requirements.txt
addict==2.4.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
idna==2.10
importlib-metadata==3.1.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///tmp/build/80754af9/jedi_1596490743326/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console==6.2.0
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148959479/work
jupyterlab-pygments==0.1.2
kaggle==1.5.10
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mmcv-full==1.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1594384075987/work
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
pycocotools==2.0.2
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
scipy==1.5.4
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205313296/work
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
timm==0.3.1
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.1+cu101
tornado==6.0.4
tqdm==4.54.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
yapf==0.30.0
zipp==3.4.0
